+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 11.4
 - PostgreSQL fix limit set_updated_at_triggers() to current schema in rosariosis.sql
 - SQL add Comment to every grade of the Main scale in rosariosis.sql & rosariosis_mysql.sql
 - Add Chrome OS to GetUserAgentOS() in UserAgent.fnc.php
+- Add 'SELECT_ONLY' to $extra param in GetStaffList.fnc.php
 
 Changes in 11.3.2
 -----------------